+++ lib/models/kasseneck_receipt.dart
@@ -10,6 +10,7 @@ import 'package:kasseneck_api/services/rksv_service.dart';
 
 import '../enums/credit_card_provider.dart';
 import '../enums/keck_payment_method.dart';
+import '../enums/qr_print_mode.dart';
 import '../enums/receipt_type.dart';
 import '../enums/voucher_action.dart';
 import '../enums/voucher_type.dart';
@@ -307,9 +308,9 @@ class KasseneckReceipt implements Comparable<KasseneckReceipt> {
 
   Future<PrintResponse> printReceiptMyPos() => KeckPrinterService.printReceiptMypos(this);
   Future printReceiptWifi() => KeckPrinterService.printReceiptWifi(this);
-  Future printReceiptBluetooth({bool qrAsImage = true}) => KeckPrinterService.printReceiptBluetooth(this, qrAsImage: qrAsImage);
+  Future printReceiptBluetooth({QrPrintMode qrMode = QrPrintMode.imageRaster}) => KeckPrinterService.printReceiptBluetooth(this, qrMode: qrMode);
 
-  Future<List<Uint8List>> getPrintBytes({required KeckPaperSize paperSize, bool qrAsImage = false}) => KeckPrinterService.getBytesFromReceipt(this, paperSize, qrAsImage: qrAsImage);
+  Future<List<Uint8List>> getPrintBytes({required KeckPaperSize paperSize, QrPrintMode qrMode = QrPrintMode.imageRaster}) => KeckPrinterService.getBytesFromReceipt(this, paperSize, qrMode: qrMode);
 
   bool get isSigFailed => !RKSVService.isSigSuccess(sig);